Added thread pool class for multithreading support

+++ include/analyzer.h
@@ -6,12 +6,12 @@
 struct AnalyzerConfig
 {
     std::string root_path;
-    int thread_count = 4;
+    size_t thread_count = 4;
 };
 
 class Analyzer
 {
-public:
+  public:
     explicit Analyzer(const AnalyzerConfig &config);
 
     // Ignores comments or empty lines.
@@ -20,9 +20,8 @@ public:
     // Ignores nothing, every single line counts.
     uint32_t getTotalLinesOfCode();
 
-private:
+  private:
     AnalyzerConfig m_Config = {};
 
-    template <typename Func>
-    uint32_t processJavaFiles(Func fn);
+    template <typename Func> uint32_t processJavaFiles(Func fn);
 };
